+++ src/modules/admin/widgets/forms/ActiveForm.php
@@ -23,18 +23,21 @@ class ActiveForm extends \davidhirtz\yii2\skeleton\widgets\bootstrap\ActiveForm
 
     public function statusField(array $options = []): ActiveField|string
     {
-        return count($statuses = $this->getStatuses()) > 1 ? $this->field($this->model, 'status', $options)->dropdownList($statuses) : '';
+        return count($statuses = $this->getStatuses()) > 1
+            ? $this->field($this->model, 'status', $options)->dropDownList($statuses)
+            : '';
     }
 
     public function typeField(array $options = []): ActiveField|string
     {
-        return count($types = $this->getTypes()) > 1 ? $this->field($this->model, 'type', $options)->dropdownList($types) : '';
+        return count($types = $this->getTypes()) > 1
+            ? $this->field($this->model, 'type', $options)->dropDownList($types)
+            : '';
     }
 
     public function descriptionField(array $options = []): ActiveField|string
     {
-        $attribute = $this->model->getI18nAttributeName('description', ArrayHelper::remove($options, 'language'));
-        return $this->field($this->model, $attribute, $options)->textarea();
+        return $this->field($this->model, 'description', $options)->textarea();
     }
 
     public function nameField(array $options = []): ActiveField|string