+++ src/modules/admin/widgets/forms/AssetActiveForm.php
@@ -19,7 +19,7 @@ class AssetActiveForm extends ActiveForm
 
     public function init(): void
     {
-        $this->fields = $this->fields ?: array_diff($this->getDefaultFieldNames(), [
+        $this->fields ??= array_diff($this->getDefaultFieldNames(), [
             'file_id',
             'entry_id',
             'section_id',