+++ src/modules/admin/widgets/forms/CategoryActiveForm.php
@@ -3,7 +3,7 @@
 namespace davidhirtz\yii2\cms\modules\admin\widgets\forms;
 
 use davidhirtz\yii2\cms\models\Category;
-use davidhirtz\yii2\cms\modules\admin\widgets\CategoryTrait;
+use davidhirtz\yii2\cms\models\CategoryCollection;
 use davidhirtz\yii2\cms\modules\ModuleTrait;
 use davidhirtz\yii2\skeleton\widgets\fontawesome\ActiveField;
 use Yii;
@@ -14,7 +14,6 @@ use yii\helpers\ArrayHelper;
  */
 class CategoryActiveForm extends ActiveForm
 {
-    use CategoryTrait;
     use ModuleTrait;
 
     public int $slugMaxLength = 20;
@@ -24,15 +23,15 @@ class CategoryActiveForm extends ActiveForm
     public function init(): void
     {
         $this->fields ??= [
-            $this->statusField(),
-            $this->parentIdField(),
-            $this->typeField(),
-            $this->nameField(),
-            $this->contentField(),
+            $this->statusField(...),
+            $this->parentIdField(...),
+            $this->typeField(...),
+            $this->nameField(...),
+            $this->contentField(...),
             '-',
-            $this->titleField(),
-            $this->descriptionField(),
-            $this->slugField(),
+            $this->titleField(...),
+            $this->descriptionField(...),
+            $this->slugField(...),
         ];
 
         parent::init();
@@ -41,7 +40,7 @@ class CategoryActiveForm extends ActiveForm
     public function parentIdField(array $options = []): ActiveField|string
     {
         if (static::getModule()->enableNestedCategories) {
-            if ($categories = static::getCategories()) {
+            if ($categories = CategoryCollection::getAll()) {
                 $attributeNames = $this->model->getI18nAttributeNames('slug');
                 $defaultOptions = ['prompt' => ['text' => '']];
 
@@ -61,7 +60,9 @@ class CategoryActiveForm extends ActiveForm
                 }
 
                 $items = Category::indentNestedTree($categories, $this->model->getI18nAttributeName('name'));
-                return $this->field($this->model, 'parent_id')->dropdownList($items, ArrayHelper::merge($defaultOptions, $options));
+                $options = ArrayHelper::merge($defaultOptions, $options);
+
+                return $this->field($this->model, 'parent_id')->dropDownList($items, $options);
             }
         }
 
@@ -98,7 +99,7 @@ class CategoryActiveForm extends ActiveForm
             $route = array_filter(array_merge($route, ['language' => $urlManager->i18nUrl || $urlManager->i18nSubdomain ? $language : null]));
 
             if (isset($route['category']) && mb_strlen((string)$route['category'], Yii::$app->charset) > $this->slugMaxLength) {
-                $route['category'] = '...' . mb_substr((string)$route['category'], -$this->slugMaxLength, $this->slugMaxLength, Yii::$app->charset);
+                $route['category'] = mb_substr((string)$route['category'], -$this->slugMaxLength, $this->slugMaxLength, Yii::$app->charset);
             }
 
             return rtrim($category->isEnabled() || !$draftHostInfo ? $urlManager->createAbsoluteUrl($route) : $urlManager->createDraftUrl($route), '/') . '/';